AUTHOR: Andy A. Lapada  Associate Professor, College of Computer Studies - Eastern Samar State University, Borongan City, Eastern Samar, 6800, Philippines lapada213@gmail.com
AbstractThis study aimed to developed and evaluate a web-based management information system for the planning department of Eastern Samar State University. A software development methodology called Rapid Application development (RAD) were used in the development of an IT solution to the existing problem encountered in the planning unit of a university. ISO 9126 software quality standard was adapted in this study to evaluate the software. Based on the result of the evaluation conducted by the experts and end-users, the system got an over-all rating of 4.51 interpreted as strongly acceptable. It is recommended that the web-based management information system must be used or adopted by the planning unit of Eastern Samar State University.