Qingjie Wang, Xuegang Chen*, Sheng Duan, Luda Wang and Hongbin Fan  College of software and communication engineering, XiangNan University, Chenzhou, 423000, China  *Corresponding author (E-mail:gxcjsj@163.com)
Abstract
This study aimed to developed and evaluate a web-based management information system for the planning department of Eastern Samar State University. A software development methodology called Rapid Application development (RAD) were used in the development of an IT solution to the existing problem encountered in the planning unit of a university. ISO 9126 software quality standard was adapted in this study to evaluate the software. Based on the result of the evaluation conducted by the experts and end-users, the system got an over-all rating of 4.51 interpreted as strongly acceptable. It is recommended that the web-based management information system must be used or adopted by the planning unit of Eastern Samar State University.