Utilization of Coconut (Cocos Nucifera) and Abaca (Musa Textilis) Fiber in Production of Alternative Concrete Hollow Blocks
Abstract
Concrete hollow blocks are vital components that are used in construction of various establishments. However, due to the various materials used in construction, which contribute to environmental pollution. As to address this issue, the researchers aim to improve the quality of concrete hollow blocks (CHB) in terms of durability/strength and water absorption. Wherein, this study used natural materials such as coconut shell, abaca fiber, and coconut fiber as additional aggregates in production of alternative CHB products. Which involves gathering of materials, followed by physical treatment, production process, along with 25 days of curing process, then data collection that includes absorption test and drop test. With respect for the results, which proven that the use of natural materials, as additional aggregates in production of CHB products is an effective approach to improve its quality in terms of durability/strength and water absorption.