ORGANIC INK FROM SWEET POTATO LEAVES INFUSED WITH GUM ARABIC AS AN ALTERNATIVE SOURCE OF INK
Abstract
In the ink industry, the solvent in the printed ink content, except for the one held by the printed material evaporates in its own environment after the printing process. Most of these solvents and organic compounds used in printing environment contain at least one carbon and hydrogen atom and have negative effects on health and environment. (Aydemir et al. 2020). This experimental study explores the viability of sweet potato leaf-infused ink supplemented with gum Arabic as a sustainable alternative to traditional inks. This research contributes to the growing demand for eco-friendly inks and offers a sustainable solution to reduce environmental hazards associated with traditional inks. The research focuses on analyzing the odor, consistency, and stability of the organic ink in comparison to conventional ballpoint ink. Chromatography tests are conducted to examine the pigment properties of the sweet potato ink. Data collection involves monitoring odor changes during ink production and assessing stability over time. Descriptive statistics are utilized to analyze the data, including the number and intensity of pigments, odor formulation, and stability alterations.
The study findings indicate that sweet potato ink with gum Arabic exhibits low odor, stable color, and strong consistency. However, the color of the sweet potato ink is not yet suitable for school use, appearing as brown to greenish brown. Future research should explore alternative plant sources for darker-colored dyes and develop suitable pens for practical applications. Increasing the amount of gum Arabic could also be investigated to further understand its impact on the ink's qualities. This research contributes to the growing demand for eco-friendly inks and offers a sustainable solution to reduce environmental hazards associated with traditional inks.