The Effectiveness of Using Gumamela And Butterfly Pea Extracted Pigment as An Alternative Ink As A Writing Tool
The usage of chemical ink has an effect on the environment because it contains hazardous materials that may lead to pollution. According to S.Wild(2016) a complicated mixture of solvent, resins, alcohol, lubricants, carbon, pigment, color, aniline, dextrin, fluorescent, and other ingredients that makes up ink. Both the environment and people may be harmed by this kind of material. Solvent based ink may produce harmful substances that may result in health issues, asthma attacks, and irritations of the skin and eyes. (Packoi Team, 2016). The goal of this study is to create an alternative ink made of organic materials by utilizing gumamela and butterfly pea flowers. This study examines the durability and quality of ink produced when applied to various paper grades measured in grams per square meter.
The study results show that alternative ink can be produced using gumamela and butterfly pea. It shows that the longevity of pigment can be retained on different types of paper. The study also proves that it is important to know what solvent should be used. This study uses alcohol and water as solvents. In future research, it is recommended to find other suitable concentrations and paper types. This research can help improve the production of alternative inks.