'THE UTILIZATION OF MUD AND COCO PEAT AS AN ALTERNATIVE SOIL IN CULTIVATING TOMATO
Abstract
Mud is a slimy or sticky mixture of dirt or other fine-grained earth materials and water. It is mostly used in aquaculture to build pathways that serve as boundaries of fish ponds. However, if disregarded, the mud that gets stuck in rivers, streams, and canals contributes to the increasing intensity of flood and hightide problems in different areas. On the other hand, coco peat is a waste material sourced from coconut shells. His has been proven to be an effective medium in cultivating plants and has long been used in agriculture. In this regard, the researchers came up with the goal to utilize the mud that is trapped in such waterways to contribute in lowering excessive flooding and, at the same time, make use of it as an additional planting medium together with the large amount of coco peat wastes.The tomato plants cultivated in pure mud, mud and coco peat mixture, and garden soil were taken care of and monitored. An analysis of variance showed that the type of medium has no significant difference in cultivating the tomato plants in terms of height, F (7, 48) = 0.876, p = 0.532. However, the conducted experiment proved that pure mud and mud and coco peat mixture can be used in cultivating tomato plants. Based on the conclusions, the study suggested that the study may be conducted to determine the effects of these media on the quality of fruits produced by the tomato plants and that the produced media may be tested in other types of plants.