The Effectiveness Of Bamboo Briquettes Made From Bayog (Dendrocalamus Merrillianus) In Hagonoy Bulacan
Abstract
This study aims to reduce the use of conventional charcoal by introducing a new fuel source that the researchers believe to be more effective and sustainable than charcoal. The new fuel source is a charcoal briquette made with bamboo specifically Bayog (Dendrocalamus merrillianus) that is native to the researchers and a binder made with cornstarch and water. 4 tables show the differences of the briquettes and the charcoal, and based on the results the briquettes proved to be much more efficient than conventional charcoal in burning time, ignition time, max heat and heat consistency. Because of the results that we have collected we can say that briquettes made from Bayog (Dendrocalamus Merrillianus) have a significant difference in comparison to conventional charcoal.