Celso, Jean Claud A. Faustino, John Aaron Y. Quitiol, Jocelyn B. Resurrecion, Kurt Venjo Santos, Jack Daniel R.
Abstract
This study indicates that Hagonoy, Bulacan, is the region most at risk for natural
disasters such as tropical storms, tsunamis, storm surges, and monsoons. In order to
address this issue, the researchers conclude that it would be helpful to investigate and
better understand the role that mangroves play in preventing natural disasters as well as
the unique environmental contribution that each mangrove has. The researchers
identified that there are seven species of mangroves in Hagonoy, Bulacan, with the help
of Edwino S. Fernando, Ph.D. for the Institute of Biology at the University of the
Philippines, Diliman. The researchers’ findings are Acrostichum aereum L.,
Rhizophora apiculata Blume, Rhizophora stylosa Griff, Acanthus ebracteatus Vahl,
Nypa fruticans Wurmb, and Sonneratia caseolaris (L.) Engl. in Engl. & Prantl, and
Avicennia species. And each species has its own unique characteristics and uses. This
study concludes that R. apiculata, R. stylosa, N. fruticans, and Avicennia Species
protect the coastline region. Moreover, the last project is "mangrove propagation".