Buensuceso, Rona D.M. Caparas, Leilla Kydie D. Carreon, Fladee Mhel C. De Lizo, Michelle Aileendenise R.
Abstract
The intention to replace portions of firewood and charcoal and thereby reduce
the cutting down of forests and carbon dioxide emissions from charcoal carbonization
for fuel purposes, efforts are being made to utilize biomass residues. In this regard, the
researchers came up with the making of alternative briquette with alternative binder.
The briquette successfully burn like the conventional wood charcoal but the burning
time is not same. A table shows the comparison of conventional wood charcoal to
alternative briquette based on the effectivity and the properties of Balangot (Typha
Domingensis Pers.) plant is suitable as the main material for the production of
briquettes. Based on the conclusions, the study suggested that the ratio of ingredients
can still be changed as it can change the effectiveness.