Arellano, Anna Leah B. Dela Cruz, Ma. Angela R. Escaran, Charisse J. Ventura, Jeanette Carla J. Victorino, Eunice C.
Abstract
Hagonoy, Bulacan is known for having low-lying areas, therefore flooding is a
major problem. Due to this, the researchers created an alternative NPK fertilizer for the
selected plants namely Okra and Eggplant to survive in a hyperhydrated condition. As
a result of this research, the alternative NPK fertilizer in Eggplant has significant effects
in terms of viability and growth improvement since it did not produce any symptoms
of root rot disease and grows 2 cm more than the unfertilized plant. In Okra, it also has
significant effects in terms of viability since it also did not develop any symptoms of
root rot disease, but the alternative NPK fertilizer has insufficient evidence to prove the
growth improvement of Okra due to the limited time spent in conducting the research.
In terms of leaf color, the alternative NPK fertilizer has significant effects in fertilized
plants since it developed a green and darker green color compared to the unfertilized
plants.