Acuña, Raizen F. Cervantes, Julia Nicole M. Danguilan, Kate Alexandra S. Sioco, Gillian B. Valdez, Jerson Patrick C.
Abstract
Utilization of crab shells is unheard of in a societal setting. Therefore, being
surrounded with bodies of water, crab shells are one of the most generated waste
products in Hagonoy, Bulacan, leading to unpleasant odors and poor environmental
aesthetics. For this reason, a remediation method is necessary in order to alleviate this
problem. Case in point is the utilization of powdered crab shell and extracted chitin as
fertilizers as previous studies have proved it to be potent in enhancing the growth of
plants due to its high nitrogen content and low C/N ratio. The preparation process
involved physical and chemical treatments, such as the removal of protein and calcium
carbonate with sodium hydroxide and hydrochloric acid, respectively. The products
were then applied every week in varying amounts as fertilizer through soil application
for a month-long duration. However, results showed that extracted chitin can be
detrimental to plant growth if no regular monitoring of pH level take place. Hence, it
causes the plant to raise its soil pH level and deteriorate. Consequently, an analysis of
variance showed that although powdered crab shell enabled easier application as
fertilizer, it had no significant effect on the growth of the tomato plant in terms of
height, but is proven efficient in increasing the growth of leaves in terms of size. Which
is why based on the conclusions of the study, the fertilizers’ testing in other types of
plants, constant pH level monitoring, and testing in different setting and condition is
therefore recommended for future studies in relation to the topic under study.